+++ jwp-basic-gradle/src/main/java/next/api/auth/LogoutController.java
@@ -1,0 +1,22 @@
+package next.api.auth;
+
+import core.web.ModelAndView;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import next.common.controller.AbstractController;
+import next.common.view.JspView;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class LogoutController extends AbstractController {
+    private static final Logger log = LoggerFactory.getLogger(LogoutController.class);
+
+    @Override
+    protected ModelAndView doGet(HttpServletRequest req, HttpServletResponse resp) {
+        HttpSession session = req.getSession();
+        session.invalidate();
+        return new ModelAndView(new JspView("redirect:/index"));
+    }
+}
