+++ jwp-basic-gradle/src/main/java/next/api/user/CreateUserController.java
@@ -1,0 +1,36 @@
+package next.api.user;
+
+import core.web.ModelAndView;
+import next.common.controller.AbstractController;
+import next.api.user.dao.UserDao;
+import next.api.user.model.User;
+import next.common.view.JspView;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.sql.SQLException;
+
+public class CreateUserController extends AbstractController {
+    private static final Logger log = LoggerFactory.getLogger(CreateUserController.class);
+
+    @Override
+    protected ModelAndView doPost(HttpServletRequest request, HttpServletResponse response) {
+        User user = new User(
+                request.getParameter("userId"),
+                request.getParameter("password"),
+                request.getParameter("name"),
+                request.getParameter("email")
+        );
+
+        UserDao userDao = new UserDao();
+        try {
+            userDao.insert(user);
+        } catch (SQLException e) {
+            log.error(e.getMessage());
+        }
+
+        return new ModelAndView(new JspView("redirect:/user/list"));
+    }
+}
