+++ jwp-basic-gradle/src/main/java/next/api/user/UpdateUserController.java
@@ -1,0 +1,39 @@
+package next.api.user;
+
+import core.web.ModelAndView;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import next.common.controller.AbstractController;
+import next.api.user.dao.UserDao;
+import next.api.user.model.User;
+import next.common.view.JspView;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.sql.SQLException;
+
+public class UpdateUserController extends AbstractController {
+    private static final Logger log = LoggerFactory.getLogger(UpdateUserController.class);
+    private final UserDao userDao = new UserDao();
+
+    @Override
+    protected ModelAndView doPost(HttpServletRequest request, HttpServletResponse response) {
+        User user = new User(
+                request.getParameter("userId"),
+                request.getParameter("password"),
+                request.getParameter("name"),
+                request.getParameter("email"));
+        log.debug("updateUser : {}", user);
+        try {
+            userDao.update(user);
+        } catch (SQLException e) {
+            log.error(e.toString());
+        }
+
+        HttpSession session = request.getSession();
+        session.setAttribute("user", user);
+        return new ModelAndView(new JspView("redirect:/user/list"));
+    }
+}
