+++ jwp-basic-gradle/src/main/java/next/api/user/model/User.java
@@ -1,0 +1,91 @@
+package next.api.user.model;
+
+public class User {
+    private String userId;
+    private String password;
+    private String name;
+    private String email;
+
+    public User(String userId, String password, String name, String email) {
+        this.userId = userId;
+        this.password = password;
+        this.name = name;
+        this.email = email;
+    }
+
+    public String getUserId() {
+        return userId;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public boolean isAuthWith(User user) {
+        return isAuthWith(user);
+    }
+
+    public boolean isAuthWith(String id, String password) {
+        if (!this.getUserId().equals(id)) {
+            return false;
+        }
+        if (!this.getPassword().equals(password)) {
+            return false;
+        }
+        return true;
+    }
+
+    public boolean isSameUser(String id) {
+        return this.getUserId().equals(id);
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((email == null) ? 0 : email.hashCode());
+        result = prime * result + ((name == null) ? 0 : name.hashCode());
+        result = prime * result + ((userId == null) ? 0 : userId.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        User other = (User) obj;
+        if (email == null) {
+            if (other.email != null)
+                return false;
+        } else if (!email.equals(other.email))
+            return false;
+        if (name == null) {
+            if (other.name != null)
+                return false;
+        } else if (!name.equals(other.name))
+            return false;
+        if (userId == null) {
+            if (other.userId != null)
+                return false;
+        } else if (!userId.equals(other.userId))
+            return false;
+        return true;
+    }
+
+    @Override
+    public String toString() {
+        return "id:" + userId + ", pw:" + password + ", name:" + name + ", email:" + email;
+    }
+}
