--- jwp-basic-gradle/src/main/java/next/controller/auth/LogoutController.java
@@ -1,22 +1,0 @@
-package next.controller.auth;
-
-import core.web.ModelAndView;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-import next.controller.AbstractController;
-import next.view.JspView;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class LogoutController extends AbstractController {
-    private static final Logger log = LoggerFactory.getLogger(LogoutController.class);
-
-    @Override
-    protected ModelAndView doGet(HttpServletRequest req, HttpServletResponse resp) {
-        HttpSession session = req.getSession();
-        session.invalidate();
-        return new ModelAndView(new JspView("redirect:/index"));
-    }
-}
