--- jwp-basic-gradle/src/main/java/next/controller/user/CreateUserController.java
@@ -1,36 +1,0 @@
-package next.controller.user;
-
-import core.web.ModelAndView;
-import next.controller.AbstractController;
-import next.dao.UserDao;
-import next.model.User;
-import next.view.JspView;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.sql.SQLException;
-
-public class CreateUserController extends AbstractController {
-    private static final Logger log = LoggerFactory.getLogger(CreateUserController.class);
-
-    @Override
-    protected ModelAndView doPost(HttpServletRequest request, HttpServletResponse response) {
-        User user = new User(
-                request.getParameter("userId"),
-                request.getParameter("password"),
-                request.getParameter("name"),
-                request.getParameter("email")
-        );
-
-        UserDao userDao = new UserDao();
-        try {
-            userDao.insert(user);
-        } catch (SQLException e) {
-            log.error(e.getMessage());
-        }
-
-        return new ModelAndView(new JspView("redirect:/user/list"));
-    }
-}
