--- jwp-basic-gradle/src/main/java/next/controller/user/UpdateUserFormController.java
@@ -1,35 +1,0 @@
-package next.controller.user;
-
-import core.web.ModelAndView;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-import next.controller.AbstractController;
-import next.model.User;
-import next.view.JspView;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class UpdateUserFormController extends AbstractController {
-    private static final long serialVersionUID = 1L;
-    private static final Logger log = LoggerFactory.getLogger(UpdateUserFormController.class);
-
-    @Override
-    protected ModelAndView doGet(HttpServletRequest request, HttpServletResponse response) {
-        HttpSession session = request.getSession();
-        Object value = session.getAttribute("user");
-        if (value == null) {
-            log.debug("Only logged-in users can access.");
-            return new ModelAndView(new JspView("redirect:/user/login.jsp"));
-        }
-        User user = (User)value;
-        if (!user.isSameUser(request.getParameter("userId"))) {
-            log.debug("자신의 계정에만 접근할 수 있습니다.");
-            return new ModelAndView(new JspView("redirect:/user/list"));
-        }
-
-        request.setAttribute("user", user);
-        return new ModelAndView(new JspView("/user/update.jsp")).addModel("user", user);
-    }
-}
