+++ jwp-basic-gradle/src/test/java/next/dao/UserDaoTest.java
@@ -1,6 +1,7 @@
 package next.dao;
 
-import next.model.User;
+import next.api.user.dao.UserDao;
+import next.api.user.model.User;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
